chore: refactoring secp256k1
