--- packages/ed25519/src/__fixtures__/ed25519-jws.json
@@ -1,52 +1,0 @@
-[
-  {
-    "id": "did:key:z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL#z6Mkk7yqnGF3YwTrLpqrW6PGsKci7dNqh1CjnvMbzrMerSeL",
-    "payload": {
-      "hello": "world"
-    },
-    "jws": {
-      "compact": "eyJhbGciOiJFZERTQSJ9.eyJoZWxsbyI6IndvcmxkIn0.VzfMtSAyy3Xd9KioZIBq2KQqbkDPmo4hC36IOe4OZ579lsMBOurUXJ2Nsi5u7TR51-F0MEvfAjBBMjizNgW1CA",
-      "detached": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..s-Tt5I-yyyhTYBIEUCzNp7VpTieFBNBxbKX-PJeWJzHRJaV_Pj4YSsPHX8slM6gXkYlBAox7oMmqZzRd6vL0Cg"
-    }
-  },
-  {
-    "id": "did:key:z6MkoBGpA9TtazNKVdoe7dgpA3G1qqq6H9Vziopp3Sz6LHWv#z6MkoBGpA9TtazNKVdoe7dgpA3G1qqq6H9Vziopp3Sz6LHWv",
-    "payload": {
-      "hello": "world"
-    },
-    "jws": {
-      "compact": "eyJhbGciOiJFZERTQSJ9.eyJoZWxsbyI6IndvcmxkIn0.jXlWaMDF2CGAuabvrfHnkiHd7iUpogfpHxSfHWsI1g1QlkI9EmUMbW9mYPTeIHqQCKHyrmvF3Cn1OnEvyTxFDA",
-      "detached": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..zJOFyoLn42nEIGgK4PjllSO55h12ozbbtXBK2DuNUBU566HA5ODhRy-SMFCTkHaCzFrwoNR8eyx7gUCMhbq6Cg"
-    }
-  },
-  {
-    "id": "did:key:z6MkquUCsHbwTmJL6bNX1rrnJdEAVL2NbUfr8KszQowjvFyo#z6MkquUCsHbwTmJL6bNX1rrnJdEAVL2NbUfr8KszQowjvFyo",
-    "payload": {
-      "hello": "world"
-    },
-    "jws": {
-      "compact": "eyJhbGciOiJFZERTQSJ9.eyJoZWxsbyI6IndvcmxkIn0.FIXKZUGQqIPqNw0TYOukflnh-_GrlSt_eV-PTj6a1YtXOPtqdIyMuM4x-pQXiKAZBs9K6Wu2dgUIzyufCQjlAw",
-      "detached": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..K9LubUX2XCzETybz_0_h4jO1Q96pNOcJh7Fb_yrOBacsob4qBHA5vH2hsL5QzcBBO6eFw5sArtzzlFHQ87ReCQ"
-    }
-  },
-  {
-    "id": "did:key:z6MkfgR1CSArpxkdoHp9XPLqMBroNK4bJmGYar8NDWbtRbX8#z6MkfgR1CSArpxkdoHp9XPLqMBroNK4bJmGYar8NDWbtRbX8",
-    "payload": {
-      "hello": "world"
-    },
-    "jws": {
-      "compact": "eyJhbGciOiJFZERTQSJ9.eyJoZWxsbyI6IndvcmxkIn0.LX1WCjW_B56bVLihjxoGmkY-6pfEx1gzaWpUJn6xb6Ss6HZSe-skFFPtSRDdbgvQ1nGEqxRFY1NF-rnrEXAABw",
-      "detached": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..DaJQ7fd7BwE2bS6_RjLQQ19XjUNxvu7RZxr1gxt2GL176qXyCYpQEhUvG5RRzOBLYtGI7xX0PT40fq3zBQdUBQ"
-    }
-  },
-  {
-    "id": "did:key:z6Mkjm2JGg2wpzyBBRLFrC2RDv7KUxQC4gAkxkvkrBK3yoXD#z6Mkjm2JGg2wpzyBBRLFrC2RDv7KUxQC4gAkxkvkrBK3yoXD",
-    "payload": {
-      "hello": "world"
-    },
-    "jws": {
-      "compact": "eyJhbGciOiJFZERTQSJ9.eyJoZWxsbyI6IndvcmxkIn0.Y3EuGJ_kgMIwQM0IuunJ5PCgroeSVTpiRq4mTA7kY-_LTfWcPQbhULrhkPxoT8nzYOyXtS926H7kNXf5PFFXBg",
-      "detached": "eyJhbGciOiJFZERTQSIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..QqChIpcto6s3JJgXrFMHAjB7_rz5TIA4IbyDzarqv21yU-3PfHR1uSpS7FGj65XVf8C7yodQRAgfk0tB-9_9Dg"
-    }
-  }
-]
