+++ packages/ed25519/src/typings.d.ts
