+++ packages/secp256k1/src/__fixtures__/README.md
@@ -1,0 +1,3 @@
+This directory is intentionally empty.
+
+Changes to fixtures can be written here, before being copied to test vectors.
