+++ packages/secp256k1/src/typings.d.ts
