+++ packages/test-vectors/src/did-core-conformance/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   ed25519: require('./ed25519'),
   x25519: require('./x25519'),
+  secp256k1: require('./secp256k1'),
 };
