+++ packages/test-vectors/src/did-core-conformance/secp256k1/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  key: require('./secp256k1.json'),
+  // jws: require('./ed25519-jws.json'),
+};
