--- packages/test-vectors/src/x25519-keypair.json
@@ -1,82 +1,0 @@
-{
-  "keypair": [
-    {
-      "seed": "0000000000000000000000000000000000000000000000000000000000000000",
-      "X25519KeyAgreementKey2019": {
-        "id": "#z6LSeu9HkTHSfLLeUs2nnzUSNedgDUevfNQgQjQC23ZCit6F",
-        "type": "X25519KeyAgreementKey2019",
-        "controller": "did:key:z6LSeu9HkTHSfLLeUs2nnzUSNedgDUevfNQgQjQC23ZCit6F",
-        "publicKeyBase58": "4Dy8E9UaZscuPUf2GLxV44RCNL7oxmEXXkgWXaug1WKV",
-        "privateKeyBase58": "11111111111111111111111111111111"
-      },
-      "JsonWebKey2020": {
-        "id": "#z6LSeu9HkTHSfLLeUs2nnzUSNedgDUevfNQgQjQC23ZCit6F",
-        "type": "JsonWebKey2020",
-        "controller": "did:key:z6LSeu9HkTHSfLLeUs2nnzUSNedgDUevfNQgQjQC23ZCit6F",
-        "publicKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "x": "L-V9o0fNYkMVKNqsX7spBzD_9oSvxM_C7ZCZX1jLO3Q"
-        },
-        "privateKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "d": "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
-          "x": "L-V9o0fNYkMVKNqsX7spBzD_9oSvxM_C7ZCZX1jLO3Q"
-        }
-      }
-    },
-    {
-      "seed": "0000000000000000000000000000000000000000000000000000000000000001",
-      "X25519KeyAgreementKey2019": {
-        "id": "#z6LStiZsmxiK4odS4Sb6JmdRFuJ6e1SYP157gtiCyJKfrYha",
-        "type": "X25519KeyAgreementKey2019",
-        "controller": "did:key:z6LStiZsmxiK4odS4Sb6JmdRFuJ6e1SYP157gtiCyJKfrYha",
-        "publicKeyBase58": "J3PiFeuSyLugy4DKn87TwK5cnruRgPtxouzXUqg99Avp",
-        "privateKeyBase58": "11111111111111111111111111111112"
-      },
-      "JsonWebKey2020": {
-        "id": "#z6LStiZsmxiK4odS4Sb6JmdRFuJ6e1SYP157gtiCyJKfrYha",
-        "type": "JsonWebKey2020",
-        "controller": "did:key:z6LStiZsmxiK4odS4Sb6JmdRFuJ6e1SYP157gtiCyJKfrYha",
-        "publicKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "x": "_TOE4TKtAqVsePRVR-5AA43HkAK5DSntkOCO7nYq5xU"
-        },
-        "privateKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "d": "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAE",
-          "x": "_TOE4TKtAqVsePRVR-5AA43HkAK5DSntkOCO7nYq5xU"
-        }
-      }
-    },
-    {
-      "seed": "0000000000000000000000000000000000000000000000000000000000000002",
-      "X25519KeyAgreementKey2019": {
-        "id": "#z6LSoMdmJz2Djah2P4L9taDmtqeJ6wwd2HhKZvNToBmvaczQ",
-        "type": "X25519KeyAgreementKey2019",
-        "controller": "did:key:z6LSoMdmJz2Djah2P4L9taDmtqeJ6wwd2HhKZvNToBmvaczQ",
-        "publicKeyBase58": "CgTbngDMe7yHHfxPMvhpaFRpFoQWKgXAgwenJj8PsFDe",
-        "privateKeyBase58": "11111111111111111111111111111113"
-      },
-      "JsonWebKey2020": {
-        "id": "#z6LSoMdmJz2Djah2P4L9taDmtqeJ6wwd2HhKZvNToBmvaczQ",
-        "type": "JsonWebKey2020",
-        "controller": "did:key:z6LSoMdmJz2Djah2P4L9taDmtqeJ6wwd2HhKZvNToBmvaczQ",
-        "publicKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "x": "rYxIwmdlrqetxTYolgXBq-qVBQCT29IYyWq9JIGgNWU"
-        },
-        "privateKeyJwk": {
-          "kty": "OKP",
-          "crv": "X25519",
-          "d": "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAI",
-          "x": "rYxIwmdlrqetxTYolgXBq-qVBQCT29IYyWq9JIGgNWU"
-        }
-      }
-    }
-  ]
-}
